+++ tests/test_osmnx.py
@@ -1,7 +1,8 @@
-"""
-OSMnx tests
------------
-"""
+################################################################################
+# test_osmnx.py
+# License: MIT, see full license in LICENSE.txt
+# Web: https://github.com/gboeing/osmnx
+################################################################################
 
 import matplotlib as mpl
 mpl.use('Agg') #use agg backend so you don't need a display on travis-ci
@@ -13,6 +14,7 @@ if os.path.exists('.temp'):
 
 import osmnx as ox
 
+# configure OSMnx
 ox.config(log_console=True, log_file=True, use_cache=True,
           data_folder='.temp/data', logs_folder='.temp/logs', 
           imgs_folder='.temp/imgs', cache_folder='.temp/cache')
@@ -20,6 +22,7 @@ ox.config(log_console=True, log_file=True, use_cache=True,
 
 def test_imports():
 
+    # test all of OSMnx's module imports
     import ast
     import datetime
     import geopandas
@@ -41,56 +44,65 @@ def test_imports():
     import time
     import unicodedata
     import warnings
-    from collections import Counter, OrderedDict
+    from collections import Counter
+    from collections import OrderedDict
     from dateutil import parser
     from descartes import PolygonPatch
-    from itertools import chain, groupby
+    from itertools import chain
+    from itertools import groupby
     from matplotlib.collections import LineCollection
     from rtree.index import Index
-    from shapely.geometry import Point, MultiPoint, LineString, MultiLineString, Polygon, MultiPolygon
     from shapely import wkt
+    from shapely.geometry import Point
+    from shapely.geometry import MultiPoint
+    from shapely.geometry import LineString
+    from shapely.geometry import MultiLineString
+    from shapely.geometry import Polygon
+    from shapely.geometry import MultiPolygon
     from shapely.ops import unary_union
 
 
 def test_logging():
 
-	import logging as lg
-
-	ox.log('test debug', level=lg.DEBUG)
-	ox.log('test info', level=lg.INFO)
-	ox.log('test warning', level=lg.WARNING)
-	ox.log('test error', level=lg.ERROR)
+    # test OSMnx's logger
+    import logging as lg
+    ox.log('test a fake debug', level=lg.DEBUG)
+    ox.log('test a fake info', level=lg.INFO)
+    ox.log('test a fake warning', level=lg.WARNING)
+    ox.log('test a fake error', level=lg.ERROR)
 
 
 def test_geometry_coords_rounding():
 
-	from shapely.geometry import Point, MultiPoint, LineString, MultiLineString, Polygon, MultiPolygon
+    # test the rounding of geometry coordinates
+    from shapely.geometry import Point, MultiPoint, LineString, MultiLineString, Polygon, MultiPolygon
 
-	precision = 3
+    precision = 3
 
-	shape1 = Point(1.123456, 2.123456)
-	shape2 = ox.round_shape_coords(shape1, precision)
+    shape1 = Point(1.123456, 2.123456)
+    shape2 = ox.round_shape_coords(shape1, precision)
 
-	shape1 = MultiPoint([(1.123456, 2.123456), (3.123456, 4.123456)])
-	shape2 = ox.round_shape_coords(shape1, precision)
+    shape1 = MultiPoint([(1.123456, 2.123456), (3.123456, 4.123456)])
+    shape2 = ox.round_shape_coords(shape1, precision)
 
-	shape1 = LineString([(1.123456, 2.123456), (3.123456, 4.123456)])
-	shape2 = ox.round_shape_coords(shape1, precision)
+    shape1 = LineString([(1.123456, 2.123456), (3.123456, 4.123456)])
+    shape2 = ox.round_shape_coords(shape1, precision)
 
-	shape1 = MultiLineString([[(1.123456, 2.123456), (3.123456, 4.123456)],
-	                          [(11.123456, 12.123456), (13.123456, 14.123456)]])
-	shape2 = ox.round_shape_coords(shape1, precision)
+    shape1 = MultiLineString([[(1.123456, 2.123456), (3.123456, 4.123456)],
+                              [(11.123456, 12.123456), (13.123456, 14.123456)]])
+    shape2 = ox.round_shape_coords(shape1, precision)
 
-	shape1 = Polygon([(1.123456, 2.123456), (3.123456, 4.123456), (6.123456, 5.123456)])
-	shape2 = ox.round_shape_coords(shape1, precision)
+    shape1 = Polygon([(1.123456, 2.123456), (3.123456, 4.123456), (6.123456, 5.123456)])
+    shape2 = ox.round_shape_coords(shape1, precision)
 
-	shape1 = MultiPolygon([Polygon([(1.123456, 2.123456), (3.123456, 4.123456), (6.123456, 5.123456)]),
-	                       Polygon([(16.123456, 15.123456), (13.123456, 14.123456), (12.123456, 11.123456)])])
-	shape2 = ox.round_shape_coords(shape1, precision)
+    shape1 = MultiPolygon([Polygon([(1.123456, 2.123456), (3.123456, 4.123456), (6.123456, 5.123456)]),
+                           Polygon([(16.123456, 15.123456), (13.123456, 14.123456), (12.123456, 11.123456)])])
+    shape2 = ox.round_shape_coords(shape1, precision)
 
 
 def test_gdf_shapefiles():
 
+    # test loading spatial boundaries, saving as shapefile, and plotting
     city = ox.gdf_from_place('Manhattan, New York City, New York, USA')
     city_projected = ox.project_gdf(city, to_crs={'init':'epsg:3395'})
     ox.save_gdf_shapefile(city_projected)
@@ -101,6 +113,7 @@ def test_gdf_shapefiles():
 
 def test_graph_from_file():
     
+    # test loading a graph from a local .osm file
     import bz2, tempfile
 
     node_id = 53098262
@@ -126,20 +139,29 @@ def test_graph_from_file():
 
 def test_network_saving_loading():
 
+    # save/load graph as shapefile and graphml file
     G = ox.graph_from_place('Piedmont, California, USA')
     G_projected = ox.project_graph(G)
     ox.save_graph_shapefile(G_projected)
     ox.save_graphml(G_projected)
     G2 = ox.load_graphml('graph.graphml')
 
+    # convert graph to node/edge GeoDataFrames and back again
     gdf_edges = ox.graph_to_gdfs(G, nodes=False, edges=True, fill_edge_geometry=False)
     gdf_nodes, gdf_edges = ox.graph_to_gdfs(G, nodes=True, edges=True, node_geometry=True, fill_edge_geometry=True)
     G3 = ox.gdfs_to_graph(gdf_nodes, gdf_edges)
 
+    # find graph nodes nearest to some set of points
+    X = gdf_nodes['x'].head()
+    Y = gdf_nodes['y'].head()
+    nn1 = ox.get_nearest_nodes(G, X, Y)
+    nn2 = ox.get_nearest_nodes(G, X, Y, method='kdtree')
+    nn3 = ox.get_nearest_nodes(G, X, Y, method='balltree')
+
 
 def test_get_network_methods():
 
-    import geopandas as gpd
+    from shapely import wkt
 
     # graph from bounding box
     north, south, east, west = 37.79, 37.78, -122.41, -122.43
@@ -159,27 +181,28 @@ def test_get_network_methods():
     places = ['Los Altos, California, USA', {'city':'Los Altos Hills', 'state':'California'}, 'Loyola, California']
     G5 = ox.graph_from_place(places, network_type='all', clean_periphery=False)
 
-    # graph from polygon (from shapefile)
-    calif = gpd.read_file('tests/input_data/ZillowNeighborhoods-CA')
-    mission_district = calif[(calif['CITY']=='San Francisco') & (calif['NAME']=='Mission')]
-    polygon = mission_district['geometry'].iloc[0]
+    # graph from polygon
+    polygon = wkt.loads('POLYGON ((-122.418083 37.754154, -122.418082 37.766028, -122.410909 37.766028, -122.410908 37.754154, -122.418083 37.754154))')
     G6 = ox.graph_from_polygon(polygon, network_type='walk')
 
     # test custom query filter
-    filt = ('["area"!~"yes"]'
-            '["highway"!~"motor|proposed|construction|abandoned|platform|raceway"]'
-            '["foot"!~"no"]'
-            '["service"!~"private"]'
-            '["access"!~"private"]')
-    G = ox.graph_from_point(location_point, network_type='walk', custom_filter=filt)
+    filtr = ('["area"!~"yes"]'
+             '["highway"!~"motor|proposed|construction|abandoned|platform|raceway"]'
+             '["foot"!~"no"]'
+             '["service"!~"private"]'
+             '["access"!~"private"]')
+    G = ox.graph_from_point(location_point, network_type='walk', custom_filter=filtr)
 
 
 def test_stats():
 
+    # create graph, add bearings, project it
     location_point = (37.791427, -122.410018)
     G = ox.graph_from_point(location_point, distance=500, distance_type='network')
     G = ox.add_edge_bearings(G)
     G_proj = ox.project_graph(G)
+
+    # calculate stats
     stats1 = ox.basic_stats(G)
     stats2 = ox.basic_stats(G, area=1000)
     stats3 = ox.basic_stats(G_proj, area=1000, clean_intersects=True, tolerance=15, circuity_dist='euclidean')
@@ -193,8 +216,10 @@ def test_plots():
     nc = ox.get_node_colors_by_attr(G2, 'osmid')
     ec = ox.get_edge_colors_by_attr(G2, 'length')
 
+    # save a plot to disk as png
     fig, ax = ox.plot_graph(G, save=True, file_format='png')
 
+    # save a plot to disk as svg
     G_simplified = ox.simplify_graph(G)
     fig, ax = ox.plot_graph(G_simplified, show=False, save=True, close=True, file_format='svg')
 
@@ -213,12 +238,13 @@ def test_plots():
 
 def test_routing_folium():
 
+    # calculate shortest path and plot as static image and leaflet web map
     import networkx as nx
     G = ox.graph_from_address('398 N. Sicily Pl., Chandler, Arizona', distance=800, network_type='drive')
     origin = (33.307792, -111.894940)
     destination = (33.312994, -111.894998)
     origin_node = ox.get_nearest_node(G, origin)
-    destination_node = ox.get_nearest_node(G, destination)
+    destination_node = ox.get_nearest_node(G, destination, method='euclidean')
     route = nx.shortest_path(G, origin_node, destination_node)
 
     attributes = ox.get_route_edge_attributes(G, route, 'length')
@@ -233,7 +259,8 @@ def test_routing_folium():
 
 def test_buildings():
 
-    gdf = ox.buildings_from_place(place='Piedmont, California, USA')
-    gdf = ox.buildings_from_address(address='San Francisco, California, USA', distance=300)
+    # download building footprints and plot them
+    gdf = ox.buildings_from_place(place='Emeryville, California, USA')
+    gdf = ox.buildings_from_address(address='600 Montgomery St, San Francisco, California, USA', distance=300)
     fig, ax = ox.plot_buildings(gdf)